+++ src/App.tsx
@@ -2,7 +2,7 @@ import './App.css';
 import { ReactElement } from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary';
-import MainPageProvider from './components/MainPageContext/MainPageProvider/MainPageProvider';
+import MainPageProvider from './components/MainPageContext/MainPageProvider';
 import MainPage from './components/MainPage/MainPage';
 
 const App = (): ReactElement => {
