+++ src/components/MainPage/interfaces.tsx
@@ -1,4 +1,4 @@
-import { SelectItemsOnPageProps } from '../../ItemsOnPage/ItemsOnPageInterfaces/interfaces';
+import { SelectItemsOnPageProps } from '../ItemsOnPage/interfaces';
 
 export interface SelectItemsOnPageChange extends SelectItemsOnPageProps {
   selectedValue: number;
