+++ src/utils/MainPage/clearFilteredResults.tsx
